go.mobile/sprite/portable: transform bugfixes and documentation
The 1/geom.Scale was simply an incorrect version of m.Inverse.
The texture scaling is to correct the fact that the pixels of
texture images were being conflated with the pxiels on the
screen. There is no inherent width/height implied by the pixel
dimensions of a texture.

LGTM=nigeltao
R=nigeltao
CC=golang-codereviews
https://golang.org/cl/156340044