mobile/audio: skip RIFF WAVE header if exists
PCM data may contain a header, we need to skip it and stream the
audio data to the underlying OpenAL player. Otherwise, we will play
the header bytes as well and the sound will click.

The future work will include auto detecting the number of channels,
the bitrate and the sample rate from the header.

Fixes golang/go#10728

Change-Id: Ie09e4c59a08207838d0c8f4c5d4d0cb08deb27cd
Reviewed-on: https://go-review.googlesource.com/9860
Reviewed-by: David Crawshaw <crawshaw@golang.org>